FInished

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -19,11 +19,17 @@ namespace CS321_W5D2_BlogAPI.Core.Services
 
         public Post Add(Post newPost)
         {
-            // TODO: Prevent users from adding to a blog that isn't theirs
-            //     Use the _userService to get the current users id.
-            //     You may have to retrieve the blog in order to check user id
-            // TODO: assign the current date to DatePublished
-            return _postRepository.Add(newPost);
+            var CurrentUser = _userService.CurrentUserId;
+            var CurrentBlog = _blogRepository.Get(newPost.BlogId);
+            if ( CurrentUser == CurrentBlog.UserId)
+            {
+                newPost.DatePublished = DateTime.Now;
+                return _postRepository.Add(newPost);
+
+            }
+            throw new Exception("False");
+           
+         
         }
 
         public Post Get(int id)
@@ -44,14 +50,19 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         public void Remove(int id)
         {
             var post = this.Get(id);
-            // TODO: prevent user from deleting from a blog that isn't theirs
-            _postRepository.Remove(id);
+            if (_userService.CurrentUserId == post.Blog.UserId) {
+                _postRepository.Remove(id);
+            }
+            throw new Exception("False");
         }
 
         public Post Update(Post updatedPost)
         {
-            // TODO: prevent user from updating a blog that isn't theirs
-            return _postRepository.Update(updatedPost);
+            if (_userService.CurrentUserId == updatedPost.Blog.UserId)
+            {
+                return _postRepository.Update(updatedPost);
+            }
+            throw new Exception("False");
         }
 
     }
